Virtiofsd: Add support for remote inotify/fsnotify events
Add remote inotify event support for virtiofsd if the guest also supports
inotify. There are two modes of operation; one where the guest supports
only remote events for all the files/directories within a exported dir
through FUSE and one where the guest supports only local events like in the
normal use case of inotify.

When the guest adds a watch virtiofsd will check if there is an inotify
instance, within virtiofsd that corresponds to the inotify instance in the
guest and if no then virtiofsd will create a new inotify instance else not.
Then it will add the watch tied to the created/existing inotify instance to
the inode of interest.

A separate thread created in virtiofsd will poll (with epoll) all the
inotify file descriptors/instances created for events. When an event is
generated virtiofsd will read it, check if the corresponding inode is still
active and if yes it will send an fsnotify notification to the guest to
generate a new event locally.

There are some race conditions still unresolved due to the nature of the
remote inotify/fsnotify implementation. Specifically, races occur when a
watch gets removed but at the same time an event is generated. This will
cause the guest to drop the event that is received from virtiofsd since the
local watch on the inode in the guest is now deleted. Since the "fsnotify"
function checks for the liveliness of the watch it will see that it is not
alive and thus will drop the event.