virtiofsd: Add the inotify related structs in virtiofsd
Add the structs and variables related to fsnotify/inotify that will be used
for the remote event notification subsystem. All the structs and variables
are added to the "fuse_session" struct since they need to be globaly
accessible by the inotify specific thread.

The use of a dedicated inotify thread is mandatory so that virtiofsd has
the resources to service guest requests but at the same time to monitor for
inotify events.

Also create two hashtables "wd_to_inode" and "inode_to_wd" which hold the
mappings between watch descriptors to inodes and vice versa. The purpose of
these two hashtables is to easily map watch descriptors to inodes when an
inotify occurs and virtiofsd needs to send a notification to the guest.

In addition, for each guest process that uses the inotify subsystem to monitor
events on an inode, virtiofsd will create a separate inotify instance as
well. This premise is helpful to distinguish separate watches that are
places on the same inode by different guest processes.

Finally, virtiofsd polls all the inotify instances (descriptors) for events
using the epoll subsystem.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>