virtiofsd: Add the functions that service the fsnotify/inotify requests
Add the passthrough functions that service the fsnotify requests. A
fsnotify request can implement three actions: 1) Add a new watch, 2)
Modify an existing watch and 3) Delete a watch from an inode.

In any case virtiofsd will first check if there is an existing inotify
instance corresponding to the guest process that sent the request. In the
case that the request refers to the deletion of a watch without an existing
inotify instance, an error is generated. Else virtiofsd will create a new
inotify instance and will then create a watch on the inode targeted by the
request.

Since inotify needs the name of the file to add a watch to it, something
that is not available directly in virtiofsd, readlinkat is used along with
the open proc descriptor of the target inode to find a "link" name to said
inode. Then virtiofsd uses the "inotify_add_watch" system call to add a
watch on the target inode (file or dir).

There is also a reference counter attached to each inotify instance created
by virtiofsd. With each addition of a watch related to an inotify instance
the reference counter is increased by one and when a watch is removed it is
decreased by one. When the counter reaches zero then the inotify instance
is deleted. This helps to prevent inotify instances being alive event after
the guest process associated with them exits.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>