virtiofsd: Add a fsnotify opcode and in and out message structs
Add a new opcode for the suppport of the fsnotify subsystem in virtiofsd.
Also add two new structs that will carry the fsnotify information between
the guest and virtiofsd.

The "in" struct, which corresponds to a fsnotify request will carry the
mask for the events the guest wants to monitor, an identifier that is
unique to the guest process that wants to monitor events on an inode and
the action which corresponds to the creation, modification and deletion of
a watch on an inode. The guest does not need to send the inode information
along with the fsnotify request since that information is already
encapsulated in the FUSE request.

The "out" struct corresponds to the notification that virtiofsd generates
when an event occurs and the guest has to be notified about it. This struct
carries the information about the inode for which the event ocurred, the
mask that was placed on the inode, the name of the file for which the event
occured (If that is available) and the cookie. This information corresponds
to the information that a process receives from the kernel when a event
occurs for a monitored inode. Since now virtiofsd is monitoring inodes for
events and not the guest kernel, it is the responsibility of virtiofsd to
send this information to the guest.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>