virtiofsd: Add the low level fuse functions for fsnotify
Add the low level fuse functions in virtiofsd corresponding to fsnotify
request and fsnotify notification.

The low level request servicing function parses the data within the
"fuse_notify_fsnotify_in" struct that is part of the fsnotify request sent
by the a guest. The data contains the mask for the watch, the unique id of
the guest process that sends the request and the action to be performed on
the watch. Then it calls the fsnotify corresponding handler within
passthrough_ll.c

The low level fsnotify notification function does the exact opposite, it
sends the event data about an inode to the guest through the notification
queue. Specifically, it sends the event mask placed on the monitored inode,
the inode number, a file name associated with the inode (if available) and
the cookie which is a number related to the inotify events.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>