virtiofsd: Add the inotify related structs in virtiofsd
Add a new struct "fuse_inotify" in "fuse_i.h" that holds the information
about the inotify instances (created by viritofsd) and their watches.
Add the struct to the "fuse_session" struct since "fuse_inotify" needs to
be globaly accessible by the inotify specific thread.

The use of a dedicated inotify thread is mandatory so that virtiofsd has
the resources to service guest requests but at the same time to monitor for
inotify events.

Also create two hashtables "wd_to_inode" and "inode_to_wd" which hold the
mappings between watch descriptors to inodes and vice versa. The purpose of
these two hashtables is to easily map watch descriptors to inodes when an
inotify event occurs and virtiofsd needs to send a notification to the guest.

In addition, for each guest process that uses the inotify subsystem to monitor
events on an inode, virtiofsd will create a separate inotify instance as
well. This is helpful to distinguish separate watches that are placed on the
same inode by different guest processes.

Finally, virtiofsd polls all the inotify instances (descriptors) for events
using the epoll subsystem.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>