virtiofsd: Add the inotify thread that monitors for events
Add the dedicated thread that monitors the inotify events. The thread is
constantly polling (with epoll) all the inotify instances (descriptors)
created by virtiofsd.

The thread is created during the initialization function of virtiofsd. If
the guest kernel wants remote inotify notification enabled and virtiofsd
supports it then virtiofsd creates the inotify dedicated thread.

When an event is pending for an inotify instance, the inotify thread
invokes the event handler which reads the event, maps the event to the
corresponding inode and then sends the information to the guest.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>