virtiofsd: Cleanup inotify thread when virtiofsd exits or reinitializes
During the main virtqueue thread cleanup, wait for the inotify thread to
exit safely during the virtiofsd exit or its reinitialization.

The inotify thread will check if the member-variable "cleanup_inotify" of
the struct "fuse_inotify" under struct "fuse_session". The "fuse_session"
struct is globaly accessible and it is tied to the life-time of virtiofsd.
Since the inotify thread is not tied to the life time of a virtqueue but on
the life time of virtiofsd, the use of the of "cleanup_inotify" member of
struct "fuse_inotify" seems a logical way to notify the inotify thread
about any impending exit or re-initialization of virtiofsd.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>