Aggregated mask for inotify and generation usage
Instead of creating an inotify instance per inotify instance that is
created by (a) guest process(es), we only create one inotify instance
in the server that corresponds to all the guest processes.

However, now for each inode we use its aggregated watch. Each time the
guest adds/removes a watch from an inode and as a result the
aggregated watch of that inode is modified, we send that aggregated
watch to the server which updates the watch on the corresponding inode
on the host.

Also we add support for the generation ID as an additional measure of
identification for the target inode. When a watch is placed on an
inode in the server, the generation of the inode is stored. When an
event occurs on the server, the remote event sent by the server to the
guest also contains the generation, so that the guest can validate that
the event belongs to the correct inode.